Move Schema to the right file
